+++ BuoyancyApi/BuoyancyApi.Application/RequiredTime/Commands/Update/UpdateRequiredTimeCommand.cs
@@ -13,7 +13,7 @@ public record UpdateRequiredTimeCommand
     public int RoleId { get; init; }
     public int SkillId { get; init; }
     public DateOnly WeekStartingMonday { get; init; }
-    public int Hours { get; init; }
+    public decimal Hours { get; init; }
 }
 
 public class UpdateRequiredTimeCommandHandler : IRequestHandler<UpdateRequiredTimeCommand>
