+++ BuoyancyApi/BuoyancyApi.Application/RequiredTime/Dtos/RequiredTimeDto.cs
@@ -10,7 +10,7 @@ public record RequiredTimeDto
     public int RoleId { get; set; } = default!;
     public int SkillId { get; set; } = default!;
     public DateOnly WeekStartingMonday { get; set; } = default!;
-    public int Hours { get; set; } = default!;
+    public decimal Hours { get; set; } = default!;
 
     public static RequiredTimeDto MapFromEntity(RequiredTimeEntity entity)
     {
