+++ BuoyancyApi/BuoyancyApi.E2eTests/Shared/Dtos/PlannedTimes/PlannedTimeDto.cs
@@ -8,5 +8,5 @@ public record PlannedTimeDto
     public int PersonId { get; init; }
     public int? NonProjectTimeTypeId { get; init; }
     public int WeekStartingMonday { get; init; }
-    public int Hours { get; init; }
+    public decimal Hours { get; init; }
 }
